Default to xterm-mode selection
This allows snippet plugins to do their magic tabbed selection of
positions to fill in a snippet. For example when you want to complete
an `if~`, snippet plugins populate the buffer at that cursor position
with a skeleton of an `if` statement and select the parts that you want
to customise yourself. This is only possible if the default selection
mode is xterm. So in this commit we carefully manage using mswin-style
selection when the user initiates it with SHIFT keys etc.

The only downside is that I haven't found a way to detect when mouse
selection starts in order to move it away from xterm selection to mswin
sselection.